fix(wi-6): CI — merge orphan test_no_channel_group into test__main, full AAA markers
Lead's CI heads-up (2026-05-20) on PR #124:
  * audit-all (scitex-dev 0.11.x stricter rules):
    - STX-TQ002 (missing AAA marker comments) on the regression
      tests.
    - PS-204 orphan-test-file —
      ``tests/.../cli_pkg/test_no_channel_group.py`` has no matching
      ``src/.../cli_pkg/no_channel_group.py`` mirror.

The regression is about ``cli_pkg/_main.py``'s LazyGroup not
registering ``channel`` and the CLI not advertising it. Mirror:
``tests/.../cli_pkg/test__main.py`` (already exists). Merged the
five regression tests into a new "WI-6 — ``sac channel`` tombstone
regression guard" section at the bottom of ``test__main.py``.

Each test carries explicit ``# Arrange / # Act / # Assert`` markers
and exactly one assertion (TQ007). The
``test_main_cli_module_imports_cleanly`` test was rewritten to bind
``mod = None`` + ``err`` in the except branch and assert on
``mod is not None`` so the AAA structure is unambiguous.

Deleted: tests/.../cli_pkg/test_no_channel_group.py (no longer needed).

Tests:
  * pytest tests/.../cli_pkg/test__main.py — 29 passed (24 pre-existing
    + 5 new regression tests).
  * pytest tests/.../cli_pkg/ — 848 passed, 3 skipped (no regressions).
  * Local audit: ``test_no_channel_group.py`` no longer in the PS-204
    list; remaining PS-204 entries are pre-existing on develop and
    deferred per the lead's note.
